Code duplication tweaks, fix some guidance on ollama models
