+++ graphcap_studio/src/components/ui/index.ts
@@ -4,3 +4,4 @@ export * from "./ImageCounter";
 export * from "./buttons";
 export * from "./status";
 export * from "../common_inference/ModelSelector";
+export * from "./dialog/TutorialDialog";
